% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circRNA_full.gtf.R
\name{circRNA_full.gtf}
\alias{circRNA_full.gtf}
\title{Generate circRNA Full GTF File}
\usage{
circRNA_full.gtf(samplepath = samplepath, ReferenceSet = ReferenceSet)
}
\arguments{
\item{samplepath}{Data of input, containing sample information, must have columns:    SampleName    FullPath (level to sample directory).}

\item{ReferenceSet}{Referenceset data of all possible full-lenghth isoforms.}
}
\value{
Generates a circRNA_full.gtf file in the quantification directory for each sample.
}
\description{
This function processes visualization data from circRNA detection tools
to generate a GTF file containing exon information for circular RNAs
with 'Full' isoform state.
}
\examples{
\dontrun{
circRNA_full.gtf(samplepath = samplepath, 
                 ReferenceSet = ReferenceSet)
}
}
