% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circRNA_only.gtf.R
\name{circRNA_only.gtf}
\alias{circRNA_only.gtf}
\title{Generate GTF File for circRNA Only Isoforms}
\usage{
circRNA_only.gtf(
  inputpathfile = "PRJNA429023/DataPathFile.txt",
  referencefile = "ReferenceIsoformFinal.txt"
)
}
\arguments{
\item{inputpathfile}{Path to the DataPathFile.txt containing sample information.}

\item{referencefile}{Path to the ReferenceIsoformFinal.txt containing reference isoform data.}
}
\value{
Generates a circRNA_only.gtf file in the quantification directory for each sample.
}
\description{
This function processes circRNA isoforms that are not present in the visualization data
(stout.list) but exist in the CIRI report. These isoforms are classified as "only" isoforms
and are processed using either reference data or genome annotation (GTF).
}
\examples{
\dontrun{
circRNA_only.gtf(inputpathfile = "PRJNA429023/DataPathFile.txt", 
                referencefile = "ReferenceIsoformFinal.txt")
}
}
