% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadLongIsoformFiles.R
\name{loadLongIsoformFiles}
\alias{loadLongIsoformFiles}
\title{Load long isoform data files from package reference directory}
\usage{
loadLongIsoformFiles(
  file1 = "FLcircAS.final.RData",
  file2 = "IsoCirc.final.RData",
  package = "AQUARIUMHB"
)
}
\arguments{
\item{file1}{Name of the first .RData file (without path). Defaults to "FLcircAS.final.RData".}

\item{file2}{Name of the second .RData file (without path). Defaults to "IsoCirc.final.RData".}

\item{package}{Name of the package. Defaults to "AQUARIUMHB".}
}
\value{
Invisible NULL. The function loads data into the environment rather than returning it.
}
\description{
Loads two specified .RData files located in the package's inst/reference directory.
}
\examples{
# Load default files from AQUARIUMHB package
loadLongIsoformFiles()

# Load custom files from another package
loadLongIsoformFiles("custom1.RData", "custom2.RData", package = "MyPackage")
}
