% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadSamplePathFile.R
\name{loadSamplePathFile}
\alias{loadSamplePathFile}
\title{Load Sample Paths Automatically}
\usage{
loadSamplePathFile()
}
\value{
A data frame with sample names and their full paths.
}
\description{
Automatically find SamplePathFile.txt in extdata, read sample names, and generate their full paths.
}
\examples{
ExampleSamplePath <- loadSamplePathFile()
}
