% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeReferenceIsoform.R
\name{MakeReferenceIsoform}
\alias{MakeReferenceIsoform}
\title{Generate Reference Isoform Data}
\usage{
MakeReferenceIsoform(SamplePath, outputfile)
}
\arguments{
\item{outputfile}{Path of the output file where results will be saved}

\item{samplepath}{Path dataframe of sample paths}
}
\value{
A data frame containing the combined and processed reference isoforms
}
\description{
This function processes circRNA isoform data from multiple sources,
combines them, and generates a comprehensive reference isoform dataset.
}
\examples{
\dontrun{
MakeReferenceIsoform(SamplePath, "ReferenceIsoformFinal.txt")
}
}
