% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Merge_circRNA.gtf.R
\name{Merge_circRNA.gtf}
\alias{Merge_circRNA.gtf}
\title{Merge and Process Circular RNA GTF Annotation Files}
\usage{
Merge_circRNA.gtf(SamplePath, output_file = "final_circRNA.datatable.txt")
}
\arguments{
\item{SamplePath}{Dataframe containing sample paths (output from fread("DataPathFile.txt")).}

\item{output_file}{Character, path for the output file (default: "final_circRNA.df.txt").}
}
\value{
Dataframe with merged circRNA isoform annotations.
}
\description{
Combines multiple circRNA GTF annotation files, standardizes column names,
processes isoform states, and generates a consolidated dataframe of circRNA isoforms.
}
\examples{
\dontrun{
SamplePath <- loadSamplePathFile()
merged_isoforms <- Merge_circRNA.gtf(SamplePath, output_file = "final_circRNA.datatable.txt")
}

}
