% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circRNA_full.gtf.R
\name{circrna_full.gtf}
\alias{circrna_full.gtf}
\title{Generate circRNA Full GTF File}
\usage{
circrna_full.gtf(
  datapathfile = "PRJNA429023/DataPathFile.txt",
  referencefile = "ReferenceIsoformFinal.txt"
)
}
\arguments{
\item{datapathfile}{Path to the DataPathFile.txt containing sample information.}

\item{referencefile}{Path to the ReferenceIsoformFinal.txt containing reference isoform data.}
}
\value{
Generates a circRNA_full.gtf file in the quantification directory for each sample.
}
\description{
This function processes visualization data from circRNA detection tools
to generate a GTF file containing exon information for circular RNAs
with 'Full' isoform state.
}
\examples{
\dontrun{
circRNA_full.gtf(datapathfile = "PRJNA429023/DataPathFile.txt", 
                 referencefile = "ReferenceIsoformFinal.txt")
}
}
